create a search directory in catalog
