+++ src/Catalog/Search/Category.php
@@ -5,14 +5,16 @@
  */
 declare(strict_types = 1);
 
-namespace Wizaplace\SDK\Catalog;
+namespace Wizaplace\SDK\Catalog\Search;
 
-final class SearchCategoryPath
+final class Category
 {
     /** @var int */
     private $id;
+
     /** @var string */
     private $name;
+
     /** @var string */
     private $slug;
