+++ src/Catalog/Search/SearchService.php
@@ -1,0 +1,31 @@
+<?php
+/**
+ * @copyright Copyright (c) Wizacha
+ * @license Proprietary
+ */
+declare(strict_types = 1);
+
+namespace Wizaplace\SDK\Catalog\Search;
+
+use Wizaplace\SDK\AbstractService;
+
+final class SearchService extends AbstractService
+{
+    public function search(string $query = '', array $filters = [], array $sorting = [], int $resultsPerPage = 12, int $page = 1): Result
+    {
+        $results = $this->client->get(
+            'catalog/search/products',
+            [
+                'query' => [
+                    'filters' => $filters,
+                    'sorting' => $sorting,
+                    'resultsPerPage' => $resultsPerPage,
+                    'page' => $page,
+                    'query' => $query,
+                ],
+            ]
+        );
+
+        return new Result($results);
+    }
+}
