--- tests/Catalog/CatalogServiceTest/testSearchOneProductByName.yml
@@ -1,25 +1,0 @@
-
--
-    request:
-        method: GET
-        url: 'http://wizaplace.loc/api/v1/catalog/search/products?resultsPerPage=12&page=1&query=Product'
-        headers:
-            Host: wizaplace.loc
-            Accept-Encoding: null
-            User-Agent: 'GuzzleHttp/6.2.1 curl/7.38.0 PHP/7.1.9'
-            VCR-index: '0'
-            Accept: null
-    response:
-        status:
-            http_version: '1.1'
-            code: '200'
-            message: OK
-        headers:
-            Date: 'Fri, 15 Sep 2017 06:57:39 GMT'
-            Server: 'Apache/2.4.10 (Debian)'
-            Cache-Control: 'no-cache, private'
-            X-Debug-Token: 36b309
-            X-Debug-Token-Link: 'http://wizaplace.loc/_profiler/36b309'
-            Content-Length: '6492'
-            Content-Type: application/json
-        body: '{"results":[{"productId":4,"name":"Product with shippings","slug":"product-with-shippings","minimumPrice":9.9,"crossedOutPrice":null,"isAvailable":true,"createdAt":1505458655,"updatedAt":1505458655,"transactionMode":"T","conditions":["N"],"declinationCount":1,"affiliateLink":null,"mainImage":null,"url":"\/special-category-dedicated-to-specific-tests\/product-with-shippings.html","subtitle":"","shortDescription":"La nouvelle g\u00e9n\u00e9ration de notre tablette Fire phare - d\u00e9sormais plus fine, plus l\u00e9g\u00e8re, dot\u00e9e d\u0027une plus longue autonomie et d\u0027un \u00e9cran am\u00e9lior\u00e9.","averageRating":null,"attributes":[],"categoryPath":[{"id":5,"name":"Special category dedicated to specific tests","slug":"special-category-dedicated-to-specific-tests"}],"companies":[{"id":3,"name":"The World Company Inc.","slug":"the-world-company-inc.","isProfessional":true,"image":null,"averageRating":null}]},{"productId":5,"name":"Product with complex attributes","slug":"product-with-complex-attributes","minimumPrice":15,"crossedOutPrice":null,"isAvailable":true,"createdAt":1505458655,"updatedAt":1505458655,"transactionMode":"T","conditions":["N"],"declinationCount":1,"affiliateLink":null,"mainImage":null,"url":"\/special-category-dedicated-to-specific-tests\/product-with-complex-attributes.html","subtitle":"","shortDescription":"","averageRating":null,"attributes":[{"attribute":{"id":3,"name":"Commentaire","type":"T","position":0,"parentId":null,"code":""},"values":[{"id":null,"attributeId":3,"name":"Commentaire #12M%M_\u00b009\u00a3*\/.?","slug":"","image":null}]},{"attribute":{"id":1,"name":"Couleur","type":"M","position":0,"parentId":null,"code":""},"values":[{"id":2,"attributeId":1,"name":"Blanc","slug":"blanc","image":null},{"id":3,"attributeId":1,"name":"Rouge","slug":"rouge","image":null}]},{"attribute":{"id":7,"name":"Compl\u00e9ment adresse","type":"T","position":0,"parentId":4,"code":""},"values":[{"id":null,"attributeId":7,"name":"Compl\u00e9ment d\u0027adresse pour la stepo","slug":"","image":null}]},{"attribute":{"id":5,"name":"Couleur secondaire","type":"M","position":0,"parentId":4,"code":""},"values":[{"id":9,"attributeId":5,"name":"Orange","slug":"orange","image":null}]},{"attribute":{"id":6,"name":"Pointure","type":"S","position":0,"parentId":4,"code":""},"values":[{"id":17,"attributeId":6,"name":"44","slug":"44","image":null}]},{"attribute":{"id":2,"name":"Taille","type":"S","position":0,"parentId":null,"code":""},"values":[{"id":5,"attributeId":2,"name":"M","slug":"m","image":null}]},{"attribute":{"id":null,"name":"Free attribute multiple","type":"F","position":0,"parentId":null,"code":""},"values":[{"id":null,"attributeId":null,"name":"r\u00e9ponse - 1 #","slug":"","image":null},{"id":null,"attributeId":null,"name":"r\u00e9ponse - 2 @","slug":"","image":null},{"id":null,"attributeId":null,"name":"4985","slug":"","image":null}]},{"attribute":{"id":null,"name":"Free attribute simple","type":"F","position":0,"parentId":null,"code":""},"values":[{"id":null,"attributeId":null,"name":"valeur simple du free attribute #12M%M_\u00b009\u00a3*\/.?","slug":"","image":null}]},{"attribute":{"id":null,"name":"Free attribute simple mais en tableau","type":"F","position":0,"parentId":null,"code":""},"values":[{"id":null,"attributeId":null,"name":"une bien belle valeur d\u00e9j\u00e0 encapsul\u00e9e","slug":"","image":null}]},{"attribute":{"id":null,"name":"Free attribute integer ?","type":"F","position":0,"parentId":null,"code":""},"values":[{"id":null,"attributeId":null,"name":"92254094","slug":"","image":null}]},{"attribute":{"id":null,"name":"Free attribute integer mais en tableau","type":"F","position":0,"parentId":null,"code":""},"values":[{"id":null,"attributeId":null,"name":"la m\u00eame histoire par ici","slug":"","image":null}]}],"categoryPath":[{"id":5,"name":"Special category dedicated to specific tests","slug":"special-category-dedicated-to-specific-tests"}],"companies":[{"id":3,"name":"The World Company Inc.","slug":"the-world-company-inc.","isProfessional":true,"image":null,"averageRating":null}]},{"productId":6,"name":"Product with valid geoloc","slug":"product-with-valid-geoloc","minimumPrice":0.5,"crossedOutPrice":null,"isAvailable":true,"createdAt":1505458655,"updatedAt":1505458655,"transactionMode":"T","conditions":["N"],"declinationCount":1,"affiliateLink":null,"mainImage":null,"url":"\/special-category-dedicated-to-specific-tests\/product-with-valid-geoloc.html","subtitle":"","shortDescription":"","averageRating":null,"attributes":[],"categoryPath":[{"id":5,"name":"Special category dedicated to specific tests","slug":"special-category-dedicated-to-specific-tests"}],"companies":[{"id":3,"name":"The World Company Inc.","slug":"the-world-company-inc.","isProfessional":true,"image":null,"averageRating":null}]}],"pagination":{"page":1,"nbResults":3,"nbPages":1,"resultsPerPage":12},"facets":[{"name":"categories","label":"Cat\u00e9gorie","values":{"5":{"label":"Special category dedicated to specific tests","count":"3","position":"0"}},"isNumeric":false},{"name":"price","label":"Prix","values":{"min":0.5,"max":15},"isNumeric":true},{"name":9,"label":"color","values":{"19":{"label":"blue","count":0,"position":0},"18":{"label":"red","count":0,"position":1},"20":{"label":"test attribute variant slug","count":0,"position":2}},"isNumeric":false},{"name":3,"label":"Commentaire","values":[],"isNumeric":false},{"name":1,"label":"Couleur","values":{"2":{"label":"Blanc","count":0,"position":0},"1":{"label":"Bleu","count":0,"position":1},"3":{"label":"Rouge","count":0,"position":2}},"isNumeric":false},{"name":7,"label":"Compl\u00e9ment adresse","values":[],"isNumeric":false},{"name":5,"label":"Couleur secondaire","values":{"8":{"label":"Jaune","count":0,"position":0},"9":{"label":"Orange","count":0,"position":1},"7":{"label":"Vert","count":0,"position":2},"10":{"label":"Violet","count":0,"position":3}},"isNumeric":false},{"name":6,"label":"Pointure","values":{"11":{"label":"38","count":0,"position":0},"12":{"label":"39","count":0,"position":1},"13":{"label":"40","count":0,"position":2},"14":{"label":"41","count":0,"position":3},"15":{"label":"42","count":0,"position":4},"16":{"label":"43","count":0,"position":5},"17":{"label":"44","count":0,"position":6}},"isNumeric":false},{"name":2,"label":"Taille","values":{"6":{"label":"L","count":0,"position":0},"5":{"label":"M","count":0,"position":1},"4":{"label":"S","count":0,"position":2}},"isNumeric":false}]}'
